+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/SyncPrefs.kt
@@ -3,6 +3,7 @@ package ac.mdiq.podcini.storage.model
 import io.github.xilinjia.krdb.types.RealmObject
 import io.github.xilinjia.krdb.types.annotations.PrimaryKey
 
+@Suppress("PropertyName")
 class SyncPrefs: RealmObject {
     @PrimaryKey
     var id: Long = 0L